refactor: remove unused enums and models, update data types in database schemas

+++ lib/src/features/dashboard/data/datasource/budget_datasource.dart
@@ -41,8 +41,8 @@ class BudgetDatasource {
       CREATE TABLE $table (
         id INTEGER PRIMARY KEY AUTOINCREMENT,
         type TEXT NOT NULL,
-        value DOUBLE NOT NULL,
-        target DOUBLE NOT NULL
+        value INTEGER NOT NULL,
+        target INTEGER NOT NULL
       )
     ''');
   }
